Update financial.py
Add the federal taxes. 
Added a list of states with no taxes. deleted the states with no states from the list of states. Then made the list of states a list of states with taxes. 
added the return statement for the function estimated_taxes. 
added the state tax conditional for the states with no taxes. 
added the state tax condition if/elif statements for all states with taxes